+++ openvdb_houdini/houdini/SOP_OpenVDB_Advect_Level_Set.cc
@@ -1,6 +1,6 @@
 ///////////////////////////////////////////////////////////////////////////
 //
-// Copyright (c) 2012-2013 DreamWorks Animation LLC
+// Copyright (c) 2012-2014 DreamWorks Animation LLC
 //
 // All rights reserved. This software is distributed under the
 // Mozilla Public License 2.0 ( http://www.mozilla.org/MPL/2.0/ )
@@ -340,7 +340,7 @@ SOP_OpenVDB_Advect_Level_Set::evalAdvectionParms(OP_Context& context, AdvectionP
     evalString(str, "lsGroup", 0, now);
     parms.mLSGroup = matchGroup(*gdp, str.toStdString());
 
-    parms.mTimeStep = evalFloat("timestep", 0, now);
+    parms.mTimeStep = static_cast<float>(evalFloat("timestep", 0, now));
 
     evalString(str, "advectSpatial", 0, now);
 
@@ -484,6 +484,6 @@ SOP_OpenVDB_Advect_Level_Set::processGrids(AdvectionParms& parms, hvdb::Interrup
     return true;
 }
 
-// Copyright (c) 2012-2013 DreamWorks Animation LLC
+// Copyright (c) 2012-2014 DreamWorks Animation LLC
 // All rights reserved. This software is distributed under the
 // Mozilla Public License 2.0 ( http://www.mozilla.org/MPL/2.0/ )
